+++ src/main/java/org/nmfw/foodietree/domain/store/openapi/LicenseResDto.java
@@ -1,4 +1,4 @@
-package org.nmfw.foodietree.domain.store.dto.resp;
+package org.nmfw.foodietree.domain.store.openapi;
 
 import java.util.List;
